use SO_KEEPALIVE to prevent disconnections under NAT
Fixes a defect when calling add_task() on idle connections
neither actually send any data, so the client became "stuck".

If a client is behind a NAT and doesn't send any data,
after some 2 hours the connection silently disappears from
NAT tables due to timeout. However, Gearman (and system TCP
stack) still thinks it's connected. So when you add a task
after a long timeout, the sender thinks that connection is open,
but NAT thinks the packet is spoofed, so it ignores it.