Add develop deps to META.json
It's several Module::Install::* modules. They are needed in the
author context. I'm not sure if there are tools which really
install the develop dependencies for us but the more metadata
the better. The problem is that you can't run perl Makefile.PL
in the author context unless the develop deps are installed.
Now in CI scripts we install them manually which is not ideal.